refactor: #1180: Moved support message models to Atlas.Client.Models project.
