+++ Atlas.Common/Notifications/NotificationSender.cs
@@ -1,7 +1,7 @@
 using Atlas.Common.ApplicationInsights;
-using Atlas.Common.Notifications.MessageModels;
 using System;
 using System.Threading.Tasks;
+using Atlas.Client.Models.SupportMessages;
 
 namespace Atlas.Common.Notifications
 {
