+++ Atlas.DonorImport.Test/DonorImportMessageSenderTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Atlas.Client.Models.SupportMessages;
 using Atlas.Common.Notifications;
 using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Logger;
@@ -13,7 +14,6 @@ using Dasync.Collections;
 using NSubstitute;
 using NSubstitute.ExceptionExtensions;
 using NUnit.Framework;
-using NUnit;
 using Atlas.DonorImport.Exceptions;
 
 namespace Atlas.DonorImport.Test
