+++ Atlas.MatchingAlgorithm/Services/DataRefresh/Notifications/DataRefreshSupportNotificationSender.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Atlas.Client.Models.SupportMessages;
 using Atlas.Common.Notifications;
 using Atlas.MatchingAlgorithm.Config;
