+++ Atlas.MultipleAlleleCodeDictionary/ExternalInterface/MacImporter.cs
@@ -4,6 +4,7 @@ using Atlas.MultipleAlleleCodeDictionary.Services.MacImport;
 using Dasync.Collections;
 using System;
 using System.Threading.Tasks;
+using Atlas.Client.Models.SupportMessages;
 using Atlas.Common.AzureStorage.TableStorage;
 using Atlas.Common.Notifications;
